chore: rename file from window.lua to super-commit.lua
